[Breaking] Skip non-published specs when "--release" option is set
When the `--release` option is set, Reffy used to crawl the published version of
each spec, falling back to the nightly version of the spec for specs that do not
have a published version.

Reffy now no longer falls back to the nightly version and rather simply ignores
non-published specs when `--release` is set.

This makes it possible to create a /TR view in Webref that only contains /TR
data, and avoids crawling ~275 specs that only have a nightly version to prepare
that view.

The /TR crawl is thus no longer "complete" in that it does not contain living
standards and other specs that do not have a published version. That said, known
tools that make use of Webref data should not need to do anything on their end
to accommodate this breaking change.