Drop legacyValue extraction logic
As reported in #1215, the `legacyValue` logic is no longer needed since CSS
Color 4 no longer uses it. The /TR version of the spec still uses it but...
the code stopped working a few months ago when we switched to function names
without surrounding `<>`, meaning that Webref data already does not have
`legacyValue` anymore.

This update drops that custom logic entirely. Code continues to concatenate
values if more than one definition is found.