Support "name" attribute convention in headings
Via #1223.

This extends the extraction of headings to also support the convention
where the heading ID is actually embedded within a child `<a name>`
anchor, as in CSS 2.1.

The numbering extraction logic was also adjusted to handle the case when a
top level heading starts with a number not followed by a `.`. Again, that is
the case in CSS 2.1, but note it is also the case for some IETF RFCs.

Similarly, appendix titles sometimes start with "Appendix" and may be followed
by a `:`. This update also extends the logic to support that convention.