Add dfns pre-processing logic for CSS 2.1 (#1226)
Via #1221.

The CSS 2.1 spec was published a long time ago and follows its own definitions
data model. This update adds a pre-processing step to create the expected markup
structure.

Note the spec follows different models at once. For instance, it also contains
definitions that follow the usual definitions data model. Or it uses `<dfn>`
elements without IDs. Or it starts with a `<span>` with an anchor inside or
starts with the anchor with a `<span>` inside.

Linking text for some of the extracted dfns remains clunky. For instance,
`<integer>` gets exported as `<integer>::definition of`. But then that is what
the spec defines...

Also, the spec sometimes defines the same term more than once, e.g.
"containing block" defined in:
https://www.w3.org/TR/CSS21/visuren.html#containing-block
https://www.w3.org/TR/CSS21/visuren.html#x3
https://www.w3.org/TR/CSS21/visudet.html#x0