Bump minimal supported version of Node.js from v14 to v18
Triggered by dependency on ReSpec (#1270). Note the dependency is only in tests
meaning that Reffy should continue to run with Node.js v14 for now. That said,
it does not seem a good idea to maintain two different minimal versions of
Node, one for the source and one for tests. Plus there would be no way to
specify that in `package.json`.

A major version of Reffy will need to be released.