Force version of Puppeteer across dependencies
This adds an `overrides` instruction to `package.json` to force use of the same
version of Puppeteer across dependencies. The version being used is the one of
our direct dependency.

This is a workaround for the parallel Puppeteer install issue, to help resume
Webref jobs execution on GitHub. Even if the install issue gets fixed, it may
be a good idea to keep that instruction to avoid downloading and installing
multiple versions of Chrome. In practice, the only other dependency that
requires Puppeteer is ReSpec, and our code simply does not use the bits of
code in ReSpec that use Puppeteer.