[css extraction] Handle subsidiary at-rules
The `@font-feature-values` at-rule can contain subsidiary at-rules, defined in
the spec as at-rules scoped to `@font-feature-values`:
https://drafts.csswg.org/css-fonts/#font-feature-values-syntax

Reffy was not expecting at-rules to be scoped and extracted them as root
at-rules, with a `for` attribute that did not pass schema validation.

With this update, Reffy no longer lists subsidiary at-rules at main at-rules,
and rather lists them as descriptors under the main at-rule they are scoped to.