Fix determination of path to the webidl2 module
Code that runs in the page needs to import the WebIDL2 parser. This is done
with a fake script URL that gets intercepted and converted to the contents of
the right webidl2 module file.

The conversion from the fake script URL to the webidl2 module file used some
dedicated logic to find the supposed `node_modules` folder that contains
installed modules. However, that logic assumed that all modules would be
installed at the same level. That is (no longer) always the case: when
different versions of the same dependency are used, some of the versions may
be installed in nested `node_modules` folders.

That typically happens in Webref: Reffy currently depends on commander v11.1.0
while another module depends on commander v11.0.0. Reffy's version of commander
gets installed under a `node_modules` subfolder within Reffy's folder as a
result (but not webidl2). This confused the conversion logic.

The code now rather uses `require.resolve` to determine the path of the webidl2
module.