Headings extraction - ignore empty IDs (#1423)
Clunky specs may wrap a heading with an ID in a section that has an empty ID,
as currently happens for one heading in the EME spec.

The code thought the presence of an `id` attribute was sufficient and extracted
the heading without id as a result. It now makes sure that it is not empty.