Make refs extraction more flexible and stable, add tests
There were not tests for the extraction of references! 😱

Adding tests revealed that some cases were not properly covered. Not a big deal
when extraction missed a few references. More problematic when extraction
starts extracting terms it should not. The latter happens when crawling some of
the AOM specs that are about to be added to browser-specs, see:
https://github.com/w3c/browser-specs/issues/1088

The extraction now also supports flat lists (used in some AOM specs and in
few older or external specs) and is more resilient to situations where
"references" appears in headings that are *not* about references per se (e.g.,
a change log that contains a section entitled "Changes to informative
references").

The code is not perfect but manages to extract a few references that were
previously missed, such as those in ECMA-402. It also works better with AOM
specs.

Further improvements are still possible. Typically, the code could be tied with
`create-outline.mjs` to scope sections more correctly.