Do not choke on PDF specs (#1435)
When given a PDF URL, Reffy would crash because Puppeteer rejects them.

This update makes Reffy (roughly) detect PDF URLs and silently skip them,
returning empty extracts instead.

The detection of PDF URLs currently relies on them ending with `.pdf`, which is
not ideal but avoids more complex code (network interception is done at the
request stage. To detect a PDF mime type, the interception would rather need to
be done when a response is received from the server).

Internally, a PDF URL is the same thing as an empty HTML document, so nothing
will be extracted. The title extraction now reuses the information from
browser-specs when it cannot extract the title, to avoid return "No title found
for X" whereas we actually have the info.