Align shortname schema with that in browser-specs
The comment does say that the definition should be the same as that in
browser-specs, but the update made to the definition in browser-specs back in
September 2022 for the shortname of the FIDO spec (`fido-v2.1`) had not been
copied to Reffy yet.

This is now needed because the webauthn-3 spec has started to link to the FIDO
spec.