[IDL extraction] Ignore asides added by Bikeshed (#1463)
Bikeshed may add definition panels next to the IDL code. These panels were not
properly ignored during the extraction, see for instance:
https://github.com/w3c/webref/blob/a9bd984c13414f17677dec8d06ba4028e4f74294/ed/idl/trusted-types.idl#L55

The update makes the code ignore asides. It also generalizes the code that was
dealing with such side info to exclude, and adds a few tests to check that
part.