[CSS extract] Fix dl case for SVG 1.1 spec (#1474)
Hasty copy-and-paste meant the code that dealt with old CSS property
definitions in a dl list (essentially, the SVG 1.1 spec) crashed.

In practice, the CSS extract of SVG 1.1 does not appear in Webref because
CSS extracts are only provided for the current specification (SVG 2), so this
bug did not have any visible consequence. Still, Reffy crashed on SVG 1.1.