Add links back to spec in idlparsed extracts
With this update, the idlparsed postprocessing module completes the parsed
structures with an `href` property that links back to the definition of the
IDL construct in the spec wherever possible.

That is not always possible. That should signal a problem with the spec. Main
reasons why linking back to definitions may not be possible:
- The spec simply does not have the definitions. That happens for a number
of enumeration values for instance.
- The spec does not use the appropriate dfn convention for IDL members. That's
the case for WebGL specs but also still a problem for some attributes and
methods in the HTML spec.
- The IDL in Webref is patched (for good reasons), and no longer matches what's
defined in the spec.
- The spec has an informative/normative hiccup: it defines the IDL as normative
but the actual dfns are in an informative section.