Add links back to spec in idlparsed extracts (#1493)
With this update, the idlparsed postprocessing module completes the parsed
structures with an `href` property that links back to the definition of the
IDL construct in the spec wherever possible.

The logic handles variations between Bikeshed and Respec, and works well with
overloaded methods in particular. It may link back an overloaded method back
to the "default" dfn when there is no more specific dfn, but that seems a good
enough approximation in practice.

Linking back is not always possible. That should signal a problem with the spec.
Main reasons why linking back to definitions may not be possible:
- The spec simply does not have the definitions. That happens for a number
of enumeration values for instance.
- The spec does not use the appropriate dfn convention for IDL members. That's
the case for WebGL specs but also still a problem for some attributes and
methods in the HTML spec.
- The IDL in Webref is patched (for good reasons), and no longer matches what's
defined in the spec.