Fix post-processing on specs that crawler skipped (#1498)
Reffy now skips specs that do not have a nightly URL, such as ISO specs. When
that happens, the result of the crawl for the spec is an empty object.

The post-processor runs modules at the spec level and at the crawl level, and
used the presence of a `crawled` property to distinguish between these levels.
It thus attempted to run modules that apply at the crawl level to the empty
spec crawl result.

The post-processor now also checks whether there is a `results` property.
It still looks at `crawled` to account for the unlikely case where we decide to
create a crawl module named `results`.