Use util.inspect() to report crawl errors (#1499)
When a spec cannot be crawled, the `error` property is set to something that
looked like a detailed error with a stack trace, but that actually skipped over
the chain of causes that led to the final `throw` clause. In the end, when a
fetch failed, the fetched URL appeared in the message but not the reason why
the fetch failed.

Using `util.inspect()` gives us a more complete error message. That's targeted
at debugging, and message may not be stable from one version of Node.js to
another, but understanding precisely what went wrong is exactly what we need in
such cases!