Adjust schema of links extracts, fix code indentation and casing
The changes made to links extraction changes the structure of the extracts and
the schema is no longer accurate as a result. That did not show up because we
don't have specific tests on links extraction for now.

This updates the schema to the new structure.

Also taking the opportunity to fix the innovative tab-and-space indentation of
the code and to use consistent casing for variable names.