Extract information on cancelability of events (#1534)
* Extract information on cancelability of events

* Adapt schema and test suite

* Also add `cancelable` to the schema of the events post-processor

The `cancelable` property is copied as-is by the post-processor, which should
be what we want.

* Events extraction: report `cancelable` mismatches to log

In theory, specs could set or unset the `cancelable` property of a given event
differently whenever they fire the event. In practice, the cancelability of an
event does not depend on the firing conditions. In other words, whenever a spec
fires a given event, its `cancelable` property should always be the same.

That assumption seems true today. To ease debugging in case that changes in the
future, this update makes the extraction code report an error to the log when it
bumps into discrepancies for the `cancelable` property.

(Of course, ideally, we would have a better mechanism to report such extraction
problems than hiding them in the log...)