Events post-processor: consolidate events that "babble"
It seems possible to end up with events that bubble for some target interfaces
but don't for other target interfaces, see discussion in:
https://github.com/w3c/webref/issues/1212

The post-processor already knew how to assign a bubbling flag per target
interface because we make a distinction between interfaces that are in a
bubbling tree and interfaces that are not.

However, the post-processor did not know how to consolidate an entry that
appears duplicated in the events extract because the underlying event bubbles
in one case but not in the other. This update makes it able to consolidate
these events (which I call "babbling" events because, hey, it's Friday
evening).

We still don't have post-processing tests but that seems to work fine on Webref
data.