Stop fetching resources from test.csswg.org (#1579)
Same update as that done in browser-specs:
https://github.com/w3c/browser-specs/pull/1348

Resources from test.csswg.org are used to add test annotations to a CSS spec.
We don't need that information (at least for now), and these resources seem to
contribute to network timeouts. This update adds them to the blocklist of
resources that get intercepted when the spec is loaded in Puppeteer.