Re-write throttling logic with queues
This re-writes the convoluted logic introduced in previous commit (and the
also convoluted logic that pre-existed) as suggested in #1582, taking
inspiration from the queue in the CG monitor:
https://github.com/w3c/cg-monitor/blob/5ceea24a1eaa7c1736d38909bcbb47df4ae297a3/lib/caching-queued-fetch.js#L68

On top of serializing requests sent to a given origin and sleeping a bit in
between these requests, the new queue controls parallelization, making sure
that the code does not load too many specs at once (I stuck to 4 in parallel as
before).

The net result is essentially the same as that achieved with the previous
commit: specs are processed as if origins had been interleaved. But the new
code is more straightforward, less verbose and more readable.

Also, the code for the throttled queue is the exact same one as the code in
browser-specs (see https://github.com/w3c/browser-specs/pull/1356). Code is
duplicated here out of simplicity and laziness to create and maintain a shared
dependency.