Extract normative algorithms defined in specs
This adds a browserlib module that creates extracts filled with information
about algorithms defined in the spec. The extracts are rather raw: they just
capture the tree structure of algorithms and copy the HTML for each step. This
makes the resulting extracts less directly useful but makes it possible to run
all sorts of analyses on them.

The structure of the extracts is very likely going to change substantively as
we learn from experience!

An algorithm extract is essentially an object with the following keys:
- `name`: The name of the algorithm, when one exists
- `href`: The URL with fragment to reach the algorithm, when one exists
- `html`: Some introductory prose for the algorithm. That prose may well
contain actual algorithmic operations, e.g.: "When invoked, run the following
 steps in parallel". href/src attributes in the HTML have absolute URLs.
- `steps`: Atomic algorithm steps.

Each step is essentially an object that follows the same structure as an
algorithm, except that it does not have a `name` and `href` keys, and may
also have the following keys:
- `operation`: Gives the name of the main operation performed by the step,
when one was identified. So far, that's only for "switch".
- `case`: Used in switch steps to identify the switch condition that triggers
the step.
- `ignored`: Ordered lists found at the step level that do no look like
algorithm steps. Or maybe they are? The lists should get reviewed: they
usually describe inputs/outputs or conditions, but they may signal parts
where the extraction logic needs to be improved. The lists are reported as
text prose.
- `additional`: Each step should contain one and only one algorithm. When
other algorithms are found at the same level, they get reported in that
property. That usually either signals that the spec could be improved
because if fails to use different list items for different steps, and/or
that the extraction logic needs to be smarter.