Add initial tests, fix detection of set of steps
The detection only looked at the first step in a list to detect an operation.
This wasn't the intent. The code now looks at all steps. This allows the code
to extract more algorithms and sets of steps than before (dropping a number of
cases that used to report `ignored`).

Now, this update also means that various lists that look like steps but aren't
actual steps get extracted and reported as algorithms. That does not affect the
existing analysis so far, but we'll need to iterate on the logic, for sure.