Use first dfn as intro anchor, namespace based on `data-algorithm-for`
When an explicitly flagged algorithm contains a `<dfn>`, that definition is now
used as the anchor to find the introductory paragraph. This is useful for cases
where the spec adds some intermediary paragraph between the actual introductory
paragraph and the steps, as in:
https://url.spec.whatwg.org/#percent-decode

The new logic is not perfect. It cannot spot the cases where the algorithm is
not explicitly linked to the steps, as in:
https://w3c.github.io/webappsec-csp/#abstract-opdef-parse-a-serialized-csp

In that case, the extract will contain two algorithms: first one with the
actual introductory paragraph and dfn, thought to be a "single-step" algorithm,
second one with the actual steps but the wrong introductory paragraph (and no
dfn), because the logic only checks the previous paragraph.

The URL spec also uses `data-algorithm-for` to create a namespace for the
algorithms. Now supported. Example:
https://url.spec.whatwg.org/#percent-decode
https://url.spec.whatwg.org/#string-percent-decode